Fix imports and remove redundant interface declaration
